feat: 설문조사 api 결과 데이터 테이블에 저장 로직 추가
